Allow get_objs_for_user to check a specific model
Previously, the model for the objects which were returned by
`get_objs_for_user` was determined based on the model that was
associated with the permission.  Because OLP allows permissions
to be assigned to any models, it is necessary to allow the model
to be overriden.  This allows anyone to retrieve a queryset for
the models which the user has a specific permission on.

Tests will follow.