Rewrite filters to use Q objects
This ends up producing smaller, more efficient queries that don't
have to abuse joins in order to get only the permissions that are
required.  There isn't much more that can be optimized within
`get_objs_for_user` because we have to make a subquery for each and
every model that the user specifies and any short-circuiting changes
would force us to split the query up into multiple, sometimes
smaller queries.  This should use the Django-generated indexes as
best as possible, so it should be relatively efficient for the
average use case.

The main advantage of using Q objects over combining multiple
querysets is that Django cannot produce non-joining SQL queries,
even when the querysets use the same base models.  This is made
considerably worse if users go and filter the querysets even
further after, as Django will force a new join each time `filter`
is called.

We are now using an internal cache on the `Permission` model where
we map permission codenames to a `Permission` object. We are also
no longer using the two-stage filter for permissions with similar
names, and instead are opting to just get the permission using a
single query.